Add Processing library
Add Processing library v. 2.1.1 to the project, to facilitate
distribution.

I guess this is alright, as they say in their
[FAQ](http://wiki.processing.org/w/FAQ) to this date:

"[...] The export libraries (also known as 'core') are released under an
LGPL license, which means they can be used as a library and included in
your project without you having to open up your code (though we
encourage people to share anyway). [...]"